requirements.txt
# Embeddings & LLM
sentence-transformers
transformers
huggingface-hub
langchain>=0.2.0
langchain-community
langchain-huggingface

# Vector store (latest SDK with GRPC)
pinecone[grpc]==3.0.0

# PDF parsing
pypdf

# App frameworks
flask
streamlit

# Environment variables
python-dotenv

# Google Sheets integration
gspread
oauth2client

# Reminders & scheduling
twilio
sendgrid
schedule
python-dateutil
